Waveguide crossings
Upon calling path_2_waveguide2() if any pf the selected paths intersects with one another, and a cell for a crossing has been configured in the waveguide XML file (and exists in the technology files), a user may insert a crossing cell automatically at the crossing location.

Modifications include minor changes in path_to_waveguide, a new file in SiEPIC utils called crossings (is this the correct location??) and a change in core waveguideGIU to include the parameter 'crossing_cell' in the search for waveguide parameters. I've included the 4crossing shell from Prof Shi (in EBeam) in the GSiP technology gds files so that the crossings functions have something to do in the generic tech by default, and modified the corresponding WAVEGUIDE.xml definition to include an example of referencing the crossing cell.