Include origin definition of the crossing cell
I figured some PDKs that include crossings, do not necessarily follow the 'having the origin in the center of the cell' rule, so I included a parameter that allows to define a different center od the cell (or crossing point for non symemetric cells).  This is  just an additional transformation applied when inserting the cell. Modified the core.WaveguideGUI to read the new parameter.