Get and Set methods for the SPICE paramters
I was thinking of options to set and get SPICE parameters in a way that is less manual than directly writing the text labels only. So I thought of extending the Component class to allow to do so.  This is with reference to the discussion # 249 in the EBeam PDK (https://github.com/SiEPIC/SiEPIC_EBeam_PDK/discussions/249).

The get method here basically is just a wrapper of the params property, but the set method checks if there is a Text cell with the SPICE parameters, changes that text  and updates the component.param property. If no TextCell is found, then it just creates one.

If this function is called outside of a PCell declaration itself, then whatever it's done is basically replaced or erased by the original cell information when it gets re-instantiated, so it would only be permanent on static cells. Still, I think it would be a useful method to for example update the waveguide parameters in the Waveguide PCell (usually this is defined in the PDK, but the SPICE parameters are defined inside the SiEPIC.utils layout calls).

This shouldn't have any compatibility issues since they are new methods. Although I did add a semicolon erasing to the params_dict method.