Update layout_waveguide2 and 3 to allow optional SBends
With this change, the sbends are optional and defined by the waveguide XML file by adding the sbends parameter to the specification. That is adding to each waveguide component:
<sbends>y</sbends>

If the said element doesn't exist as a waveguide parameter, the flag 'sbends' in  layout_waveguide3 gets set as False and passed to the layout_waveguide2 method.