Update get_technology_by_name
Update layer reading to cehck for subgrouping in the Layer Deck sepcification, following the lyp specification for doing so (creating new grou_members inside a group_member inside a group_member ....).  [in https://www.klayout.de/lyp_format.html]

Nested groups in the layer deck can with this be used to allow extra categorization. For example:
![example](https://user-images.githubusercontent.com/14344419/182823036-41f05540-803c-4dd9-a07f-333e5c448118.png)